Rename "models" folder to "modules"
