+++ deck.gl/examples/pure-js/angular/src/app/modules/home/components/chart/chart.component.scss
