+++ deck.gl/examples/pure-js/angular/src/app/modules/home/components/toggle/toggle.component.html
