+++ deck.gl/examples/pure-js/angular/src/app/modules/home/home-routing.module.ts
