+++ deck.gl/examples/pure-js/angular/src/app/modules/home/home.component.ts
