+++ deck.gl/examples/pure-js/angular/src/app/modules/home/layers/buildings-layer.ts
@@ -1,12 +1,15 @@
-import { CartoBQTilerLayer } from "@deck.gl/carto";
 import { Injectable } from "@angular/core";
-import { Layer } from "../../../layers/layer";
 import { Subject } from "rxjs";
+
+import { CartoBQTilerLayer } from "@deck.gl/carto";
+
+import { Layer } from "../../../models/layer";
 
 @Injectable()
 export class BuildingsLayer extends Layer {
 
   id = 'BUILDINGS_LAYER';
+  visible = false;
   viewportLoaded = new Subject();
 
   async getLayer() {
@@ -14,6 +17,7 @@ export class BuildingsLayer extends Layer {
       id: this.id,
       data: 'cartobq.maps.msft_buildings',
       binary: true,
+      visible: this.visible,
       pointRadiusUnits: 'pixels',
       getFillColor: [240, 142, 240],
       getFileColor: [240, 142, 240],
