+++ deck.gl/examples/pure-js/angular/src/app/modules/home/views/sidebar/sidebar.component.html
