+++ deck.gl/examples/pure-js/angular/src/app/modules/shared/components/map/map.component.html
