+++ deck.gl/examples/pure-js/angular/src/app/modules/shared/shared.module.ts
